+++ man-roxygen/plotGraph.R
@@ -1,1 +1,2 @@
+#' @param width,height Passed to \code{\link{visNetwork}}.
 #' @return \code{plotGraph} returns the result of \code{\link{visNetwork}}.